chore: move go code to internal packages
These packages are not part of a public API, make that clear following the [Internal packages](https://go.dev/doc/go1.4#internalpackages) convention.
