+++ reference-converter/internal/output/output_test.go
@@ -7,8 +7,8 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/nginxinc/nginx-directive-reference/reference-converter/output"
-	"github.com/nginxinc/nginx-directive-reference/reference-converter/parse"
+	"github.com/nginxinc/nginx-directive-reference/reference-converter/internal/output"
+	"github.com/nginxinc/nginx-directive-reference/reference-converter/internal/parse"
 	"github.com/stretchr/testify/require"
 )
