+++ reference-converter/internal/parse/code_test.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/nginxinc/nginx-directive-reference/reference-converter/parse"
+	"github.com/nginxinc/nginx-directive-reference/reference-converter/internal/parse"
 	"github.com/stretchr/testify/require"
 )
