+++ reference-converter/internal/parse/list.go
