+++ reference-converter/internal/parse/testdata/module.xml
