+++ reference-converter/internal/parse/testdata/not-a-module.xml
