+++ reference-converter/internal/parse/utils_test.go
@@ -8,7 +8,7 @@ import (
 	"testing"
 	"text/template"
 
-	"github.com/nginxinc/nginx-directive-reference/reference-converter/tarball"
+	"github.com/nginxinc/nginx-directive-reference/reference-converter/internal/tarball"
 	"github.com/stretchr/testify/require"
 )
